do not block gc if there is no changefeed
